requirements.txt
streamlit
pandas
scikit-learn
fpdf2
google-generativeai
python-dotenv
